+++ CHANGELOG.md
@@ -20,10 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - this parameter is mandatory
 - Update serverless app to use `aws_region` in default scan
   - ⚠ This _optional_ parameter is not yet supported by the scanner lib (which always defaults to the default region of the lambda)
+- Export scan result summary as OpenMetrics/Prometheus metrics (see `-m` or `--as-metrics` flag in CLI).
 
 ### Changed
 
 - Update AWS sdk dependencies.
+- Remove the out-file CLI option
 
 ## [0.0.3] - 2022-06-10